Separate registry integration into a feature
Change-type: patch
