+++ src/features/registry/certs.ts
@@ -6,8 +6,8 @@ import {
 	TOKEN_AUTH_CERT_KID,
 	TOKEN_AUTH_CERT_PUB,
 	TOKEN_AUTH_JWT_ALGO,
-} from './config';
-import { b64decode } from './utils';
+} from '../../lib/config';
+import { b64decode } from '../../lib/utils';
 
 const allowedAlgorithms = [
 	'HS256',
