+++ src/features/registry/registry.ts
@@ -1,7 +1,6 @@
 // Implements the server part of: https://docs.docker.com/registry/spec/auth/token/
 // Reference: https://docs.docker.com/registry/spec/auth/jwt/
 
-import * as BasicAuth from 'basic-auth';
 import type { Request, RequestHandler } from 'express';
 import * as jsonwebtoken from 'jsonwebtoken';
 import * as _ from 'lodash';
@@ -10,16 +9,15 @@ import * as uuid from 'uuid';
 
 import { sbvrUtils, permissions, errors } from '@balena/pinejs';
 
-import { retrieveAPIKey } from '../infra/auth/api-keys';
-import { captureException, handleHttpErrors } from '../infra/error-handling';
+import { captureException, handleHttpErrors } from '../../infra/error-handling';
 
-import { registryAuth as CERT } from '../lib/certs';
+import { registryAuth as CERT } from './certs';
 import {
 	AUTH_RESINOS_REGISTRY_CODE,
 	REGISTRY2_HOST,
 	TOKEN_AUTH_BUILDER_TOKEN,
-} from '../lib/config';
-import type { User as DbUser } from '../models';
+} from '../../lib/config';
+import type { User as DbUser } from '../../models';
 
 const { UnauthorizedError } = errors;
 const { api } = sbvrUtils;
@@ -54,26 +52,6 @@ export interface Access {
 	actions: string[];
 }
 type Scope = [Access['type'], Access['name'], Access['actions']];
-
-// Resolves permissions and populates req.user object, in case an api key is used
-// in the password field of a basic authentication header
-export const basicApiKeyAuthenticate: RequestHandler = async (
-	req,
-	_res,
-	next,
-) => {
-	const creds = BasicAuth.parse(req.headers['authorization']!);
-	if (creds) {
-		req.params.subject = creds.name;
-		req.params.apikey = creds.pass;
-	}
-	try {
-		await retrieveAPIKey(req);
-		next();
-	} catch (err) {
-		next(err);
-	}
-};
 
 const parseScope = (req: Request, scope: string): Scope | undefined => {
 	try {
